separate `App.tsx` from `main.tsx`
