+++ apps/frontend/src/main.tsx
@@ -1,13 +1,6 @@
-import React, { useState } from 'react'
+import React from 'react'
 import ReactDOM from 'react-dom/client'
-import { RouterProvider, createRouter } from '@tanstack/react-router'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { httpBatchLink } from '@trpc/client'
-
-import { trpc } from '@/api/trpcClient'
-import { routeTree } from './routeTree.gen'
-import DefaultNotFound from './components/DefaultNotFound'
-import './index.css'
+import { App } from './App'
 
 if ('serviceWorker' in navigator) {
     window.addEventListener('load', async () => {
@@ -17,38 +10,6 @@ if ('serviceWorker' in navigator) {
             console.error('Service Worker registration failed:', error)
         }
     })
-}
-
-const router = createRouter({
-    defaultNotFoundComponent: DefaultNotFound,
-    routeTree,
-})
-
-declare module '@tanstack/react-router' {
-    interface Register {
-        router: typeof router
-    }
-}
-
-function App() {
-    const [queryClient] = useState(() => new QueryClient())
-
-    const [trpcClient] = useState(() =>
-        trpc.createClient({
-            links: [
-                httpBatchLink({
-                    url: 'http://localhost:3000',
-                }),
-            ],
-        })
-    )
-    return (
-        <trpc.Provider client={trpcClient} queryClient={queryClient}>
-            <QueryClientProvider client={queryClient}>
-                <RouterProvider router={router} />
-            </QueryClientProvider>
-        </trpc.Provider>
-    )
 }
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
